Full settings interface for 4.5
